i accidentally changed the wrong file lol

+++ api/main_endpoints/routes/Advertisement.js
@@ -5,7 +5,10 @@ const {
   decodeToken,
   checkIfTokenSent,
 } = require('../util/token-functions.js');
+const logger = require('../../util/logger');
 const Advertisement = require('../models/Advertisement');
+const AuditLog = require('../models/AuditLog.js');
+const AuditLogActions = require('../util/auditLogActions.js');
 
 router.get('/', async (req, res) => {
   const count = await Advertisement.countDocuments();
@@ -39,18 +42,35 @@ router.post('/createAdvertisement', async (req, res) => {
   } else if (!await decodeToken(req)) {
     return res.sendStatus(UNAUTHORIZED);
   }
+
+  const user = await decodeToken(req)
+  if (!user) {
+    return res.sendStatus(UNAUTHORIZED)
+  }
+
   const newAd = new Advertisement({
     message: req.body.message,
     expireDate: req.body.expireDate
   });
 
-  Advertisement.create(newAd)
-    .then((post) => {
-      return res.json(post);
-    })
-    .catch(
-      (error) => res.sendStatus(BAD_REQUEST)
-    );
+  try {
+    const createdAd = await Advertisement.create(newAd);
+    await AuditLog.create({
+      userId: user._id,
+      action: AuditLogActions.CREATE_AD,
+      details: {
+        createdBy: user.email,
+        message: createdAd.message,
+        expireDate: createdAd.expireDate,
+        advertisementId: createdAd._id
+      }
+    }).catch(logger.error)
+
+    res.status(OK).send(createdAd)
+  } catch (error) {
+    logger.error('Error creating ad:', error);
+    res.sendStatus(BAD_REQUEST)
+  }
 });
 
 router.post('/deleteAdvertisement', async (req, res) => {
@@ -59,17 +79,43 @@ router.post('/deleteAdvertisement', async (req, res) => {
   } else if (!await decodeToken(req)) {
     return res.sendStatus(UNAUTHORIZED);
   }
-  Advertisement.deleteOne({ _id: req.body._id })
-    .then(result => {
-      if (result.n < 1) {
-        res.sendStatus(NOT_FOUND);
-      } else {
-        res.sendStatus(OK);
+
+  const user = await decodeToken(req)
+  if (!user) {
+    return res.sendStatus(UNAUTHORIZED)
+  }
+
+  try {
+    const adToDelete = await Advertisement.findById(req.body._id)
+
+    if (!adToDelete) {
+      return res.sendStatus(NOT_FOUND)
+    }
+
+    const deleteResult = await Advertisement.deleteOne({_id: req.body._id})
+
+    if(deleteResult.deletedCount < 1) {
+      return res.sendStatus(NOT_FOUND)
+    }
+
+    AuditLog.create({
+      userId: user._id,
+      action: AuditLogActions.DELETE_AD,
+      details: {
+        deletedBy: user.email,
+        deleted_ad: {
+          id: adToDelete._id,
+          message: adToDelete.message,
+          deletedAt: new Date()
+        }
       }
-    })
-    .catch(() => {
-      res.sendStatus(BAD_REQUEST);
-    });
+    }).catch(logger.error)
+
+    res.sendStatus(OK)
+  } catch (error) {
+    logger.error('Error deleting ad:', error)
+    res.sendStatus(BAD_REQUEST)
+  }
 });
 
 module.exports = router;
